Refactor: IssueHeader폴더 생성
